Internal (#7)
* Refactoring to improve enforcing component boundaries

Tom Hombergs had a blog post (https://reflectoring.io/java-components-clean-boundaries/) that I thought was a good idea. Be explicit with an internal package for classes that are private to that component.  Kotlin has the internal class which will also be used so classes are private to the component.  Why do internal then? Testability. I can use ArchUnit to enforce classes in the internal packages are not used inappropriately.

I like the idea because:
* guides developer to use dependency inversion principle
* enforceable with ArchUnit
* we can use sub-packages to organize our code which Java's package-private visibility doesn't support

Note: renamed the kotlin src directory name to java because there are times we still need to intermix Kotlin and Java due to missing support.  In this case, Kotlin does not support Package annotations so I had to use Java to do that.

* Updated to explain why InternalPackage and a ArchUnit test was added to enforce component boundaries

Co-authored-by: dustin <dustinshimono@gmail.com>